Switch on travis's caching features
Activated both `cargo` and arbitrary directory. Strictly,
the definition of the `cargo` cache type just takes the
whole `target` dir from the route, which will happen to
pick up OpenSSL's static libs as well, but since that's
not actually an output of cargo, it seems better to be
explicit. This way we don't have to worry about the case
that travis gets more targetted caching rules.

This is the start of investigating #1027

+++ ci/run-docker.sh
@@ -145,20 +145,25 @@ case $TARGET in
     ;;
 esac
 
-mkdir -p target/$TARGET/openssl
 install=`pwd`/target/$TARGET/openssl/openssl-install
-out=`pwd`/target/$TARGET/openssl/openssl-$OPENSSL_VERS.tar.gz
-curl -o $out https://www.openssl.org/source/openssl-$OPENSSL_VERS.tar.gz
-sha256sum $out > $out.sha256
-test $OPENSSL_SHA256 = `cut -d ' ' -f 1 $out.sha256`
 
-tar xf $out -C target/$TARGET/openssl
-(cd target/$TARGET/openssl/openssl-$OPENSSL_VERS && \
- CC=$OPENSSL_CC \
- AR=$OPENSSL_AR \
- $SETARCH ./Configure --prefix=$install no-dso $OPENSSL_OS $OPENSSL_CFLAGS -fPIC && \
- make -j4 && \
- make install)
+if [ ! -e "$install"]; then
+  mkdir -p target/$TARGET/openssl
+  out=`pwd`/target/$TARGET/openssl/openssl-$OPENSSL_VERS.tar.gz
+  curl -o $out https://www.openssl.org/source/openssl-$OPENSSL_VERS.tar.gz
+  sha256sum $out > $out.sha256
+  test $OPENSSL_SHA256 = `cut -d ' ' -f 1 $out.sha256`
+
+  tar xf $out -C target/$TARGET/openssl
+  (cd target/$TARGET/openssl/openssl-$OPENSSL_VERS && \
+   CC=$OPENSSL_CC \
+   AR=$OPENSSL_AR \
+   $SETARCH ./Configure --prefix=$install no-dso $OPENSSL_OS $OPENSSL_CFLAGS -fPIC && \
+   make -j4 && \
+   make install)
+else
+  echo 'Using cached OpenSSL static libs'
+fi
 
 # Variables to the openssl-sys crate to link statically against the OpenSSL we
 # just compiled above
